# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Household Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface InviteCode {
  code: string;
}

export interface HouseholdInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
}

export interface JoinHouseholdInput {
  /**
     * @minLength 4
     * @maxLength 32
     */
  code: string;
}

export interface ProjectInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @maxLength 1000 */
  description?: string;
}

export interface TaskInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title: string;
  /** @maxLength 1000 */
  description?: string;
  dueDate?: string;
  projectId?: string;
  assigneeIds?: string[];
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export interface TaskUpdate {
  status?: TaskUpdateStatus;
  /**
     * @minLength 1
     * @maxLength 160
     */
  title?: string;
  /** @nullable */
  dueDate?: string | null;
}

export type HouseholdSummaryRole = typeof HouseholdSummaryRole[keyof typeof HouseholdSummaryRole];


export const HouseholdSummaryRole = {
  owner: 'owner',
  member: 'member',
} as const;

export interface HouseholdSummary {
  id: string;
  name: string;
  role: HouseholdSummaryRole;
  memberCount: number;
  openTaskCount: number;
}

export interface HouseholdCreated {
  household: HouseholdSummary;
  joinCode: string;
}

export type MemberRole = typeof MemberRole[keyof typeof MemberRole];


export const MemberRole = {
  owner: 'owner',
  member: 'member',
} as const;

export interface Member {
  id: string;
  displayName: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: MemberRole;
}

export interface Project {
  id: string;
  name: string;
  /** @nullable */
  description: string | null;
  taskCount: number;
  completedTaskCount: number;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export interface Task {
  id: string;
  title: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  dueDate: string | null;
  status: TaskStatus;
  /** @nullable */
  projectId: string | null;
  assignees: Member[];
}

export type HouseholdDashboardRole = typeof HouseholdDashboardRole[keyof typeof HouseholdDashboardRole];


export const HouseholdDashboardRole = {
  owner: 'owner',
  member: 'member',
} as const;

export interface HouseholdDashboard {
  id: string;
  name: string;
  role: HouseholdDashboardRole;
  members: Member[];
  projects: Project[];
  tasks: Task[];
}

export interface ProjectDetail {
  project: Project;
  members: Member[];
  tasks: Task[];
}